Major revision to code
Simplified the optimisation algorithm due to removal of the positive constraint on the reduction factors. 
Currently code is being modified to use a generalised cost function instead of epidemic size.
Fixed the noise in the solution of  reduction factors. Noise was caused by singularity in the log function when final size was equal to group size.
Currently code is not usable.